Implement continued numbering
This adds support for the text:continue-numbering and text:continue-list
attributes. The text:continue-numbering attribute specifies whether to
continue the immediately preceding list if it has the same style.

http://docs.oasis-open.org/office/v1.2/os/OpenDocument-v1.2-os-part1.html#attribute-text_continue-numbering

The text:continue-list attribute specifes the unique identifier of a list
to continue numbering from in the document. These must also have the same
style.

http://docs.oasis-open.org/office/v1.2/os/OpenDocument-v1.2-os-part1.html#attribute-text_continue-list

https://tools.oasis-open.org/issues/browse/OFFICE-3558