Fix wrongly displayed style information of selected images.
DirectFormattingController reads only the styling on text nodes
within the selected range. This causes wrong style information
to be communicated via signalling when the selection is an image.

This commit changes this behavior to read the style info of all
character/text elements and text nodes in the selection.

Fixes the issue where all direct and paragraph style buttons
are toggled to 'on'.