Ensure various places always have a valid root element
There is a small edge between the canvas border and the office:text
element that the user can click in. Doing this returns null for the root
element (as it is the container that contains the office:text element),
causing an unhandled null reference exception to occur in some
circumstances.

Fixes #664.