Cache file data within the loaded Zip
The zip code has been refactored to ensure only a single GET request is
needed by the browser to load a document. This makes webodf slightly more
robust when loading documents as it avoids problems with the prior HEAD
request returning different file sizes, or even being unsupported by the
browser/server completely.

The current Zip implementation attempts to lazily load parts of the zip
file during decompression which is a good concept to preserve to allow
memory usage to be reduced in future. For now however, robustness of
file loading is more important.