Prevent editor toolbar being too small, at least for left-aligned elements
If there are only floating elements, the Dojo toolbar shrinks to 0 height.

To prevent this for left-aligned elements, do not set them as
float:left. (Also reorder generation to keep current tool ordering)

And for elements floating to the right, clear any floating for the toolbar element
with "clear:both" on the pseudo :after element.